Breaking Updates: - Remove dump() from the Heap interface and implementations
The original idea was for the default iterator on Sorted objects
to iterate the object in sorted order. When order did not matter, dump()
could be used.

In practice, there exists objects that are considered sorted in usage but
not in implementation. Heaps are an example of this; to return elements in
sorted order, extra work is required. If this is the behavior of the
default iterator, then this work is done implicitly, and contexts where the
object is not known to be sorted pay the penalty.

This update allows an object's default iterator to return all elements
in an arbitrary order unless otherwise stated / expected by the object
(such as a List). Objects should instead provide a separate method
(such as Heap.sorted()) for in-order traversal of their elements.

Updates:
- Add Tree interface
- Add SortedTree interface
- Add AATree implementation